+++ app/assets/javascripts/discourse/package.json
@@ -97,7 +97,7 @@
     "ember-load-initializers": "^3.0.1",
     "ember-modifier": "^4.2.0",
     "ember-on-resize-modifier": "^2.0.2",
-    "ember-qunit": "^8.1.1",
+    "ember-qunit": "^9.0.1",
     "ember-source": "~5.12.0",
     "ember-template-imports": "^4.2.0",
     "ember-test-selectors": "^7.0.0",
