+++ app/assets/javascripts/discourse/tests/setup-tests.js
@@ -10,7 +10,6 @@ import {
   setApplication,
   setResolver,
 } from "@ember/test-helpers";
-import { addModuleExcludeMatcher } from "ember-cli-test-loader/test-support/index";
 import $ from "jquery";
 import MessageBus from "message-bus-client";
 import QUnit from "qunit";
@@ -356,32 +355,6 @@ export default function setupTests(config) {
 
   const hasPluginJs = !!document.querySelector("script[data-discourse-plugin]");
   const hasThemeJs = !!document.querySelector("script[data-theme-id]");
-
-  const shouldLoadModule = (name) => {
-    if (!/\-test/.test(name)) {
-      return false;
-    }
-
-    const isPlugin = name.match(/\/plugins\//);
-    const isTheme = name.match(/\/theme-\d+\//);
-    const isCore = !isPlugin && !isTheme;
-    const pluginName = name.match(/\/plugins\/([\w-]+)\//)?.[1];
-
-    const loadCore = target === "core" || target === "all";
-    const loadAllPlugins = target === "plugins" || target === "all";
-
-    if (hasThemeJs) {
-      return isTheme;
-    } else if (isCore && !loadCore) {
-      return false;
-    } else if (isPlugin && !(loadAllPlugins || pluginName === target)) {
-      return false;
-    }
-
-    return true;
-  };
-
-  addModuleExcludeMatcher((name) => !shouldLoadModule(name));
 
   // forces 0 as duration for all jquery animations
   $.fx.off = true;
